requirements.txt
# MemeScope Dependencies
# CSE 434/534 - Miami University

# Core
anthropic>=0.25.0
Pillow>=10.0.0
requests>=2.31.0

# OCR
easyocr>=1.7.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13

# Fine-tuning (CSE 534 graduate component)
torch>=2.0.0
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.0
bitsandbytes>=0.43.0
datasets>=2.18.0
accelerate>=0.29.0

# Demo interface
gradio>=4.0.0
